--HG--
extra : convert_revision : 7fc610625ed566ff67ef3526fa27198b4bf63eeb

+++ pidgin/plugins/spellchk.c
@@ -667,7 +667,7 @@ spellchk_new_attach(PurpleConversation *conv)
 	return;
 }
 
-static int buf_get_line(char *ibuf, char **buf, int *position, int len)
+static int buf_get_line(char *ibuf, char **buf, int *position, gsize len)
 {
 	int pos = *position;
 	int spos = pos;
